requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.27.0
google-generativeai==0.3.2
pytesseract==0.3.10
easyocr==1.7.1
numpy==1.24.3
opencv-python==4.8.1.78
pdf2image==1.16.3
aiohttp==3.9.1

